# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "quant-agent-backend",
    "description": "app API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/v1/auth/change-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Change Password",
        "description": "Change password for authenticated user.",
        "operationId": "change_password_api_v1_auth_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/backtest": {
      "post": {
        "tags": [
          "backtest"
        ],
        "summary": "Submit Backtest",
        "description": "Submit a backtest for execution.",
        "operationId": "submit_backtest_api_v1_backtest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/{backtest_id}": {
      "get": {
        "tags": [
          "backtest"
        ],
        "summary": "Get Backtest Result",
        "description": "Get backtest result by ID.",
        "operationId": "get_backtest_result_api_v1_backtest__backtest_id__get",
        "parameters": [
          {
            "name": "backtest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Backtest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/{backtest_id}/stream": {
      "get": {
        "tags": [
          "backtest"
        ],
        "summary": "Stream Backtest",
        "description": "Stream backtest progress events via SSE.",
        "operationId": "stream_backtest_api_v1_backtest__backtest_id__stream_get",
        "parameters": [
          {
            "name": "backtest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Backtest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/{backtest_id}/simulation": {
      "post": {
        "tags": [
          "backtest"
        ],
        "summary": "Submit Simulation",
        "description": "Submit simulation stub — jqcli simulation wiring deferred.",
        "operationId": "submit_simulation_api_v1_backtest__backtest_id__simulation_post",
        "parameters": [
          {
            "name": "backtest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Backtest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestSimulationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/{backtest_id}/abort": {
      "post": {
        "tags": [
          "backtest"
        ],
        "summary": "Abort Backtest",
        "description": "Abort a running backtest.",
        "operationId": "abort_backtest_api_v1_backtest__backtest_id__abort_post",
        "parameters": [
          {
            "name": "backtest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Backtest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestAbortResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "AuthResponse"
      },
      "BacktestAbortResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "BacktestAbortResponse"
      },
...
      "BacktestMetricsResponse": {
        "properties": {
          "annual_return": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annual Return"
          },
          "sharpe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharpe"
          },
          "max_drawdown": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Drawdown"
          },
          "volatility": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volatility"
          },
          "win_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Win Rate"
          },
          "total_return": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Return"
          },
          "raw": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw"
          }
        },
        "type": "object",
        "title": "BacktestMetricsResponse"
      },
      "BacktestParamsInput": {
        "properties": {
          "start_date": {
            "type": "string",
            "title": "Start Date",
            "default": "2020-01-01"
          },
          "end_date": {
            "type": "string",
            "title": "End Date",
            "default": "2024-12-31"
          },
          "initial_capital": {
            "type": "number",
            "title": "Initial Capital",
            "default": 100000.0
          },
          "frequency": {
            "type": "string",
            "title": "Frequency",
            "default": "day"
          },
          "benchmark": {
            "type": "string",
            "title": "Benchmark",
            "default": "000300.XSHG"
          }
        },
        "type": "object",
        "title": "BacktestParamsInput"
      },
      "BacktestResultResponse": {
        "properties": {
          "backtest_id": {
            "type": "string",
            "title": "Backtest Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BacktestMetricsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "performance": {
            "items": {
              "$ref": "#/components/schemas/PerformancePointResponse"
            },
            "type": "array",
            "title": "Performance"
          },
          "trades": {
            "items": {
              "$ref": "#/components/schemas/TradeDayGroupResponse"
            },
            "type": "array",
            "title": "Trades"
          },
          "holdings": {
            "items": {
              "$ref": "#/components/schemas/HoldingDayGroupResponse"
            },
            "type": "array",
            "title": "Holdings"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "backtest_id",
          "status"
        ],
        "title": "BacktestResultResponse"
      },
      "BacktestSimulationResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "simulation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Simulation Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "submitted"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "BacktestSimulationResponse",
        "description": "Response for simulation submission."
      },
      "BacktestSubmitRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 65536,
            "title": "Code"
          },
          "thread_id": {
            "type": "string",
            "format": "uuid",
            "title": "Thread Id"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          },
          "params": {
            "$ref": "#/components/schemas/BacktestParamsInput"
          }
        },
        "type": "object",
        "required": [
          "code",
          "thread_id"
        ],
        "title": "BacktestSubmitRequest"
      },
      "BacktestSubmitResponse": {
        "properties": {
          "backtest_id": {
            "type": "string",
            "title": "Backtest Id"
          }
        },
        "type": "object",
        "required": [
          "backtest_id"
        ],
        "title": "BacktestSubmitResponse"
      },
...
      "ChangePasswordRequest": {
        "properties": {
          "old_password": {
            "type": "string",
            "title": "Old Password"
          },
          "new_password": {
            "type": "string",
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "old_password",
          "new_password"
        ],
        "title": "ChangePasswordRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HoldingDayGroupResponse": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "holdings": {
            "items": {
              "$ref": "#/components/schemas/HoldingRecordResponse"
            },
            "type": "array",
            "title": "Holdings"
          },
          "summary": {
            "$ref": "#/components/schemas/HoldingDaySummaryResponse"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "HoldingDayGroupResponse"
      },
      "HoldingDaySummaryResponse": {
        "properties": {
          "total_assets": {
            "type": "number",
            "title": "Total Assets",
            "default": 0
          },
          "cash": {
            "type": "number",
            "title": "Cash",
            "default": 0
          },
          "total_market_value": {
            "type": "number",
            "title": "Total Market Value",
            "default": 0
          }
        },
        "type": "object",
        "title": "HoldingDaySummaryResponse"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "PerformancePointResponse": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "strategy": {
            "type": "number",
            "title": "Strategy",
            "default": 0
          },
          "relative": {
            "type": "number",
            "title": "Relative",
            "default": 0
          },
          "benchmark": {
            "type": "number",
            "title": "Benchmark",
            "default": 0
          },
          "position_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Pct"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "PerformancePointResponse"
      },
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "full_name"
        ],
        "title": "RegisterRequest"
      },
...
      "TradeDayGroupResponse": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "trades": {
            "items": {
              "$ref": "#/components/schemas/TradeRecordResponse"
            },
            "type": "array",
            "title": "Trades"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "TradeDayGroupResponse"
      },
      "TradeRecordResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol",
            "default": ""
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "side": {
            "type": "string",
            "title": "Side",
            "default": ""
          },
          "quantity": {
            "type": "number",
            "title": "Quantity",
            "default": 0
          },
          "price": {
            "type": "number",
            "title": "Price",
            "default": 0
          }
        },
        "type": "object",
        "title": "TradeRecordResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "JWT issued by /api/v1/auth/login, set as the access_token cookie. Configure the value in Postman's Cookies modal for the host above."
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local dev server"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ]
}